+++ httpie/output/lexers/http.py
@@ -1,4 +1,68 @@
+import re
 import pygments
+
+
+RE_STATUS_LINE = re.compile(r'(\d{3})( +)(.+)')
+
+STATUS_TYPES = {
+    '1': pygments.token.Number.HTTP.INFO,
+    '2': pygments.token.Number.HTTP.OK,
+    '3': pygments.token.Number.HTTP.REDIRECT,
+    '4': pygments.token.Number.HTTP.CLIENT_ERR,
+    '5': pygments.token.Number.HTTP.SERVER_ERR,
+}
+
+RESPONSE_TYPES = {
+    'GET': pygments.token.Name.Function.HTTP.GET,
+    'HEAD': pygments.token.Name.Function.HTTP.HEAD,
+    'POST': pygments.token.Name.Function.HTTP.POST,
+    'PUT': pygments.token.Name.Function.HTTP.PUT,
+    'PATCH': pygments.token.Name.Function.HTTP.PATCH,
+    'DELETE': pygments.token.Name.Function.HTTP.DELETE,
+}
+
+
+def precise(lexer, precise_token, parent_token):
+    # Due to a pygments bug*, custom tokens will look bad
+    # on outside styles. Until it is fixed on upstream, we'll
+    # convey whether the client is using pie style or not
+    # through precise option and return more precise tokens
+    # depending on it's value.
+    #
+    # [0]: https://github.com/pygments/pygments/issues/1986
+    if precise_token is None or not lexer.options.get("precise"):
+        return parent_token
+    else:
+        return precise_token
+
+
+def http_response_type(lexer, match, ctx):
+    status_match = RE_STATUS_LINE.match(match.group())
+    if status_match is None:
+        return None
+
+    status_code, text, reason = status_match.groups()
+    status_type = precise(
+        lexer,
+        STATUS_TYPES.get(status_code[0]),
+        pygments.token.Number
+    )
+
+    groups = pygments.lexer.bygroups(
+        status_type,
+        pygments.token.Text,
+        status_type
+    )
+    yield from groups(lexer, status_match, ctx)
+
+
+def request_method(lexer, match, ctx):
+    response_type = precise(
+        lexer,
+        RESPONSE_TYPES.get(match.group()),
+        pygments.token.Name.Function
+    )
+    yield match.start(), response_type, match.group()
 
 
 class SimplifiedHTTPLexer(pygments.lexer.RegexLexer):
@@ -18,7 +82,7 @@ class SimplifiedHTTPLexer(pygments.lexer.RegexLexer):
             # Request-Line
             (r'([A-Z]+)( +)([^ ]+)( +)(HTTP)(/)(\d+\.\d+)',
              pygments.lexer.bygroups(
-                 pygments.token.Name.Function,
+                 request_method,
                  pygments.token.Text,
                  pygments.token.Name.Namespace,
                  pygments.token.Text,
@@ -27,15 +91,13 @@ class SimplifiedHTTPLexer(pygments.lexer.RegexLexer):
                  pygments.token.Number
              )),
             # Response Status-Line
-            (r'(HTTP)(/)(\d+\.\d+)( +)(\d{3})( +)(.+)',
+            (r'(HTTP)(/)(\d+\.\d+)( +)(.+)',
              pygments.lexer.bygroups(
                  pygments.token.Keyword.Reserved,  # 'HTTP'
                  pygments.token.Operator,  # '/'
                  pygments.token.Number,  # Version
                  pygments.token.Text,
-                 pygments.token.Number,  # Status code
-                 pygments.token.Text,
-                 pygments.token.Name.Exception,  # Reason
+                 http_response_type,  # Status code and Reason
              )),
             # Header
             (r'(.*?)( *)(:)( *)(.+)', pygments.lexer.bygroups(